requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
google==3.0.0
google-auth==2.9.1
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
PyJWT==2.4.0
requests==2.27.1
SQLAlchemy==1.4.39
